unpacked-zips/qr_service_pro/requirements.txt
fastapi
uvicorn[standard]
jinja2
itsdangerous
qrcode
Pillow
stripe
sqlalchemy
alembic
python-multipart
pyjwt
reportlab
python-dotenv
